pyproject.toml
[project]
name = "streamlit-article"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.20.0",
    "matplotlib>=3.10.0",
    "openai>=1.63.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
